@@ -6,7 +6,7 @@ repository: https://github.com/hosain-mohamed/curved_splash_screen
 
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.17.6 <3.0.0"
   flutter: ">=1.17.0 <2.0.0"
 
 dependencies:
